# src/main/resources/static/openapi/transaction-service-openapi.yaml
openapi: 3.0.3
info:
  title: Transaction Service API
  description: API REST para gestionar transacciones bancarias
  version: 1.0.0
servers:
  - url: http://localhost:8082/
paths:
  /report/customers/{customerId}/balances:
    get:
      summary: Obtener resumen de saldos promedio diarios del mes en curso
      operationId: generateMonthlyBalanceSummary
      description: Devuelve el saldo promedio diario del mes actual para todos los productos del cliente.
      parameters:
        - name: customerId
          in: path
          required: true
          description: Identificador único del cliente
          schema:
            type: string
            example: "c123"
      responses:
        '200':
          description: Lista de saldos promedio diarios por producto
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BalanceSummaryResponse'

  /report/ReportAvailableBalance/{productId}:
    get:
      summary: Consultar saldo disponible de un producto bancario o tarjeta de crédito
      operationId: generateReportAvailableBalance
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Saldo disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailableBalanceResponse'

  /report/ReportProductSummary/{customerId}:
    get:
      summary: Resumen consolidado de productos del cliente
      operationId: generateReportProductSummary
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Productos disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductConsolidatedSummaryResponse'

  /report/ReportProductSummaryGeneral/{customerId}:
    get:
      summary: Resumen consolidado de productos del cliente
      operationId: generateReportProductSummaryGeneral
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
        - name: start
          in: query
          required: true
          description: Fecha de inicio del periodo (formato ISO 8601)
          schema:
            type: string
            format: date-time
            example: "2025-06-01T00:00:00"
        - name: end
          in: query
          required: true
          description: Fecha de fin del periodo (formato ISO 8601)
          schema:
            type: string
            format: date-time
            example: "2025-06-30T23:59:59"
      responses:
        '200':
          description: Productos disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductGeneralSummaryResponse'

  /report/customers/{customerId}/commissions:
    get:
      summary: Obtener reporte de comisiones por producto
      operationId: generateCommissionReports
      description: Genera un resumen de comisiones cobradas por producto de un cliente en un periodo de tiempo.
      parameters:
        - name: customerId
          in: path
          required: true
          description: Identificador único del cliente
          schema:
            type: string
            example: "c123"
        - name: start
          in: query
          required: true
          description: Fecha de inicio del periodo (formato ISO 8601)
          schema:
            type: string
            format: date-time
            example: "2025-06-01T00:00:00"
        - name: end
          in: query
          required: true
          description: Fecha de fin del periodo (formato ISO 8601)
          schema:
            type: string
            format: date-time
            example: "2025-06-30T23:59:59"
      responses:
        '200':
          description: Lista de comisiones cobradas por producto
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommissionReportResponse'

  /transactions:
    get:
      summary: Listar todas las transacciones
      operationId: getAllTransactions
      responses:
        '200':
          description: Lista de transacciones
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
    post:
      summary: Registrar una nueva transacción
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionBody'
      responses:
        '200':
          description: Transacción registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'

  /transactions/{id}:
    get:
      summary: Obtener transacción por ID
      operationId: getTransactionById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Detalle de la transacción
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
    put:
      summary: Actualizar transacción por ID
      operationId: updateTransaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionBody'
      responses:
        '200':
          description: Transacción actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
    delete:
      summary: Eliminar transacción por ID
      operationId: deleteTransaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transacción eliminada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'

  /transactions/by-product/{productId}:
    get:
      summary: Listar todas las transacciones de un producto bancario
      operationId: getTransactionsByProduct
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de transacciones por producto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'

  /transactions/applyMonthlyTasks:
    post:
      summary: Simular operaciones mensuales
      operationId: applyMonthlyTasks
      responses:
        '200':
          description: Tareas mensuales completadas correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'

  /transactions/debit-cards:
    post:
      summary: Procesa un pago o retiro usando tarjeta de débito
      operationId: processCardTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionBody'
      responses:
        '200':
          description: Transacción registrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'

components:
  schemas:
    TransactionBody:
      type: object
      properties:
        sourceProductId:
          type: string
          example: prod123
        targetProductId:
          type: string
          example: prod123
        type:
          type: string
          example: "[DEPOSIT, WITHDRAWAL, PAYMENT, TRANSFER]"
        amount:
          type: number
          format: decimal
          example: 250.0
        dateTime:
          type: string
          format: date-time
          example: 2024-05-20T10:30:00

    TransactionResponse:
      type: object
      properties:
        id:
          type: string
          example: 663032b0ac82a12a8445a9d1
        sourceProductId:
          type: string
          example: prod123
        targetProductId:
          type: string
          example: prod123
        type:
          type: string
          example: "[DEPOSIT, WITHDRAWAL, PAYMENT, TRANSFER]"
        amount:
          type: number
          format: decimal
          example: 250.0
        transactionFee:
          type: number
          format: decimal
          example: 15.0
        dateTime:
          type: string
          format: date-time
          example: 2024-05-20T10:30:00

    TemplateResponse:
      type: object
      description: Respuesta procesada
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: Operación exitosa
        transactions:
          type: array
          description: Lista de transacciones
          items:
            $ref: '#/components/schemas/TransactionResponse'

    AvailableBalanceResponse:
      type: object
      description: Respuesta con saldo disponible
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: Reporte generado correctamente
        balance:
          type: object
          properties:
            productId:
              type: string
              example: 663018e0ac82a12a8445a9b0
            availableBalance:
              type: number
              format: decimal
              example: 2450.75
            productCategory:
              type: string
              example: "[CREDIT, BANK]"

    BalanceSummaryResponse:
      type: object
      description: Respuesta con saldo promedio diario
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: Reporte generado correctamente
        balanceSummary:
          type: object
          properties:
            customerId:
              type: string
              description: Identificador único del cliente
            month:
              type: string
              description: Mes al que corresponde el resumen (formato yyyy-MM)
              example: "2025-06"
            products:
              type: array
              items:
                $ref: '#/components/schemas/ProductBalanceSummary'

    ProductConsolidatedSummaryResponse:
      type: object
      description: Resumen consolidado de productos del cliente
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: Resumen generado correctamente
        summary:
          type: object
          properties:
            customerId:
              type: string
              description: Identificador único del cliente
              example: "123456"
            products:
              type: array
              description: Lista de productos financieros del cliente
              items:
                $ref: '#/components/schemas/BankProductDTO'

    ProductGeneralSummaryResponse:
      type: object
      description: Resumen consolidado de productos con actividad en el rango de fechas especificado
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: Reporte generado correctamente
        summary:
          type: object
          properties:
            customerId:
              type: string
              description: Identificador único del cliente
              example: "123456"
            start:
              type: string
              format: date-time
              description: Fecha y hora de inicio del reporte
              example: "2025-06-01T00:00:00Z"
            end:
              type: string
              format: date-time
              description: Fecha y hora de fin del reporte
              example: "2025-06-09T23:59:59Z"
            products:
              type: array
              description: Lista de productos con actividad en el intervalo
              items:
                $ref: '#/components/schemas/BankProductDTO'

    ProductBalanceSummary:
      type: object
      properties:
        productId:
          type: string
          description: Identificador único del producto bancario
        productType:
          type: string
          description: Tipo de producto financiero
        averageDailyBalance:
          type: number
          format: decimal
          description: Saldo promedio diario del mes
          example: 5230.75

    CommissionReportResponse:
      type: object
      description: Respuesta con saldo promedio diario
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: Reporte generado correctamente
        commissionReport:
          type: object
          description: Resumen de comisiones cobradas por producto durante un periodo específico.
          properties:
            productId:
              type: string
              description: Identificador único del producto (cuenta o crédito) desde el cual se originó la transacción con comisión.
              example: "prod-001"
            productType:
              type: string
              description: Tipo de producto financiero asociado.
              example: "SAVINGS"
            totalTransactions:
              type: integer
              description: Número total de transacciones con comisión registradas para el producto.
              example: 12
            totalFee:
              type: number
              format: decimal
              description: Suma total de las comisiones cobradas sobre el producto en el periodo consultado.
              example: 45.75
            currency:
              type: string
              description: Moneda en la que se registraron las comisiones.
              example: "PEN"

    BankProductDTO:
      type: object
      properties:
        id:
          type: string
          description: ID del producto
          example: id1234
        customerId:
          type: string
          description: ID del cliente asociado al producto
          example: cust1234
        type:
          type: string
          description: Tipo de producto bancario
          example: "[ SAVINGS, CURRENT, FIXED_TERM, CREDIT ]"
        status:
          type: string
          description: Estado de producto bancario
          example: "[ ACTIVE, BLOCKED_AVG_BALANCE ]"
        name:
          type: string
          description: Nombre descriptivo del producto
          example: Cuenta Ahorro Soles
        balance:
          type: number
          format: decimal
          description: Saldo actual del producto
          example: 1000.50
        holders:
          type: array
          description: Titulares de la cuenta empresarial
          items:
            type: string
          example: [ "12345678", "87654321" ]
        signers:
          type: array
          description: Firmantes autorizados
          items:
            type: string
          example: [ "99887766" ]
        details:
          type: object
          description: Detalles del producto como JSON genérico
          example:
            maintenanceFee: 0
            monthlyLimit: 2